% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_vision.R
\name{gpt_vision}
\alias{gpt_vision}
\title{Interact with OpenAI's GPT model for image analysis}
\usage{
gpt_vision(
  image_path,
  prompt = "What is in this image?",
  max_tokens = 1000,
  model = "gpt4o"
)
}
\arguments{
\item{image_path}{The path to the image file that you want to analyze.}

\item{prompt}{An optional prompt to guide the analysis. The default prompt is "What is in this image?".}

\item{max_tokens}{The maximum number of tokens to generate in the response. Defaults to 1000.}

\item{model}{The model you chose to use. gpt4o by default.}
}
\value{
Returns the textual interpretation of the image from the GPT model. If no valid response is received,
returns `NULL`.
}
\description{
This function sends an image to OpenAI's GPT model (specifically a vision-enabled version)
and returns the model's textual interpretation of the image. The image is encoded in Base64 format
and included in the request along with an optional prompt.
}
\examples{
# Assuming you have set your OPENAI_API_KEY in your environment,
# and you have an image at "path/to/your/image.jpg"
result <- gpt_vision("path/to/your/image.jpg")
cat(result)

}
