% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ollama_linux.R
\name{install_ollama}
\alias{install_ollama}
\title{Install Ollama on a Linux System}
\usage{
install_ollama()
}
\value{
Invisible NULL. The function is primarily used for its side effects
(installing Ollama) and does not return a value.
}
\description{
This function automates the installation of Ollama on Linux systems by running
the appropriate shell commands. It uses `system()` to execute a shell script
from the Ollama website, and then enables and starts the Ollama service using
`systemctl`. Note that this function requires the user to have sudo privileges
to run successfully. It is designed to run on Linux; attempts to run on other
operating systems will result in an error.
}
\details{
The function first checks if the operating system is Linux. If so, it
proceeds to run the installation commands. If the operating system is not Linux,
the function will stop and print a message indicating that Ollama installation
is only supported on Linux, providing guidance for macOS and Windows users.
}
\examples{
\dontrun{
install_ollama()
}
}
