% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama_install_model.R
\name{ollama_install_model}
\alias{ollama_install_model}
\title{Install a Model with Ollama}
\usage{
ollama_install_model(model_name)
}
\arguments{
\item{model_name}{A character string specifying the name of the model to install.}
}
\value{
The function does not return a value. It executes a system command to install the specified model.
}
\description{
This function constructs and executes a system command to install a model using `ollama pull`.
}
\examples{
\dontrun{
  ollama_install_model("mistral")
}

}
