% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama_prompt.R
\name{ollama_prompt}
\alias{ollama_prompt}
\title{Call the ollama API}
\usage{
ollama_prompt(prompt, model = "mistral", format = NULL, print_result = TRUE)
}
\arguments{
\item{prompt}{The prompt to send to the ollama API.}

\item{model}{The model to use, default is "mistral".}

\item{format}{The format of the response, default is NULL.}

\item{print_result}{Whether to print the result, default is TRUE.}
}
\value{
The response content from the API.
}
\description{
This function makes a POST request to the ollama API with the specified prompt,
model, and format. It optionally prints the response based on the `print_result` parameter.
}
\examples{
ollama_prompt("Why is the sky blue?", model = "llama2", print_result = TRUE)
}
